+++ internal/domain/dropper_config_repository.go
@@ -2,7 +2,7 @@ package domain
 
 type DropperConfigRepository interface {
 	FindById(id string) (*DropperConfig, error)
-	FindByUserId(userId int) (*DropperConfig, error)
+	FindByUserId(userId int64) (*DropperConfig, error)
 	Save(config *DropperConfig) error
 	Delete(config *DropperConfig) error
 }
